Feat/adjust renamings (#137)
* rename to sanity.config

* rename to sanity.queries
