+++ components/CardEquipment/CardEquipment.tsx
@@ -4,7 +4,7 @@ import { MdPhotoCamera } from 'react-icons/md';
 import BlockContent from '@sanity/block-content-to-react';
 import styles from './CardEquipment.module.scss';
 import { useNextSanityImage } from 'next-sanity-image';
-import { sanityConfig } from 'lib/config';
+import { sanityConfig } from 'lib/sanity.config';
 import { Equipment } from 'types/News';
 
 const CardEquipment: FC<Equipment> = ({ body, cover, onClick, title }): JSX.Element => {
