+++ components/MemberCard/MemberCard.tsx
@@ -1,5 +1,5 @@
 import styles from './MemberCard.module.scss';
-import { sanityConfig } from 'lib/config';
+import { sanityConfig } from 'lib/sanity.config';
 import { useNextSanityImage } from 'next-sanity-image';
 import Image from 'next/image';
 import { FC } from 'react';
