+++ components/SanityImage/SanityImage.jsx
@@ -1,5 +1,5 @@
 import Image from 'next/image';
-import { sanityConfig } from 'lib/config';
+import { sanityConfig } from 'lib/sanity.config';
 import { useNextSanityImage } from 'next-sanity-image';
 import React from 'react';
