+++ pages/_app.tsx
@@ -2,7 +2,7 @@ import 'styles/globals.css';
 import 'styles/offcanvas.css';
 import 'styles/Header.css';
 import { sanityClient } from 'lib/sanity.client';
-import { rrssQuery } from 'lib/queries';
+import { rrssQuery } from 'lib/sanity.queries';
 import { useEffect, useState } from 'react';
 import { RrssContext, IRrss } from 'components/context';
 import * as ga from 'lib/ga';
